files

+++ src/main/java/org/example/LessonSeven/ArithmeticOperations.java
@@ -1,19 +1,19 @@
 package org.example.LessonSeven;
 
 public class ArithmeticOperations {
-    public static int add(int a, int b) {
-        return a + b;
+    public static double add(double a, double b) {
+        return (a + b);
     }
-    public static int subtract(int a, int b) {
-        return a - b;
+    public static double subtract(double a, double b) {
+        return (a - b);
     }
-    public static double multiply(int a, int b) {
-        return a * b;
+    public static double multiply(double a, double b) {
+        return (a * b);
     }
-    public static double division(int a, int  b) {
+    public static double division(double a, double  b) {
         if (b == 0) {
             throw new IllegalArgumentException("На ноль делить нельзя");
         }
-        return (double) a / b;
+        return (a / b);
     }
 }
